requirements.txt
What did I use to develop the programs?
First version:
- Anaconda  
- Spyder 5.3.1
- Python 3.10
Future updates:
- PyCharm

# Required packages:
# Name  Version
pandas~=1.5.1
Pillow~=9.2.0